nolibc: declare most common types and syscalls
The syscalls used in the preinit have been declared according to
their respective man pages. A few types used by these syscalls were
declared as well (mostly stdint types and those used by struct stat).

A few helper functions normally provided by the libc are provided as
well, for example FD_ZERO() or makedev(). These ones are not necessarily
complete however. For example makedev() is limited to 4096 majors.

A few basic str* and mem* functions are provided, particularly memcpy()
which is used by libgcc.

All functions are declared static. This ensures that gcc can inline them
when it notices they're used only once, and that at the same time it's
possible to grab a reference to them in a pointer. The rare ones which
require to be exported are marked __attribute__((weak)) instead. This
ensures it still works if included multiple times.

If NOLIBC_IGNORE_ERRNO is defined, the errno-specific code is not built,
saving a few bytes per syscall for programs that make no use of it. If
only a few syscall need errno, it makes sense to still build with this
set and replace these syscalls by calls to the raw versions (sys_*) which
all return -errno instead of -1.