nolibc: implement support for i386 and x86_64
The code is almost the same between these two archs. syscalls 0 to 5
are implemented. i386 only uses int 0x80. The code is implemented by
mostly relying on gcc to correctly place the registers around the
syscall, so that we minimize register allocation. This is critical
on i386 which has only a few, and in general each register swap means
a larger executable in the end. Register constraints do not work when
using so many registers, so instead we use the "register asm(reg)"
construct which is reliable.

Since the syscall number is passed in eax (or rax) and the result is
retrieved from the same register, the same local variable _num is used
to pass the syscall number and to retrieve the syscall return value.