nolibc: add support for ARM EABI in ARM and Thumb2 modes
It is almost the same as the x86 version except that the syscall number
is passed via its own register (r7) and that the result is in r0 which
also happens to be the one used for the first argument. All arguments
are passed as registers. The syscall instruction used here is "svc #0".
No registers except lr are clobbered. The stack is 8-byte aligned.

It's worth noting that it's up to _start() to switch to Thumb mode by
setting pc to +1, as the kernel always calls the executable in 32-bit
mode.

This was tested on armv5, armv7, armv7t. All little endian, but
nothing seems endian-dependant.