nolibc: implement support for MIPS O32
This ABI is very tricky. First the CPU uses delayed slots so it's easy
to misread the code. Second, only 4 arguments may be passed as registers
(a0..a3) and last ones have to be pushed into the stack. Third, eventhough
the 4 first arguments are not passed via the stack, the caller has to
provision room for them in the stack so that the callee can safely write
them if it needs so. Missing this causes quite random bugs which are
tricky to troubleshoot causing argv[] to be corrupted when the program
becomes more complex. Fourth, a lot of registers are clobbered. Fifth,
the syscall return is made via two registers (v0 and v1). The first one
contains either the syscall return value or errno, and the second one
is 0 in case of success or non-zero in case of error.

It's important to look closely at the kernel side code (scall32-o32.S)
to spot all corner cases.

This code now works fine on an AR9331 (mips32).