nolibc: implement support for aarch64
This was by far the easiest one to implement. The code is very similar to
ARM except that a few instructions changed and the asm syntax changed to
drop '%' and '$' (easier to read for x86 people). The stack is 16-byte
aligned. No registers are clobbered. Some syscall numbers are not exported
by default unless a few macros are defined before including unistd.h :

  #define __ARCH_WANT_SYSCALL_NO_AT
  #define __ARCH_WANT_SYSCALL_NO_FLAGS
  #define __ARCH_WANT_SYSCALL_DEPRECATED