nolibc: add support for __builtin_strlen()
This is convenient with many situations where there are some text
strings passed as function arguments which are then measured in
length for allocation for example. Instead of computing the string
size at run time we can often compute it at build time using
__builtin_strlen(). Unfortunately this doesn't work inside a
function on its arguments, even if it's marked inline, so we have
to rename the current strlen function and replace it with a macro
which does the same.

We don't want to use it on non-constant strings because gcc's version
is larger than the current one. Similarly, while __builtin_memset()
is seldom useful, most of the times it's larger.