nolibc: declare MAX_ERRNO
This one is convenient when checking pointers returned by the system.
With a single test a pointer can be cast to unsigned long and compared
with (unsigned)-MAX_ERRNO to see if it's between -MAX_ERRNO and -1,
which normally doesn't happen with valid pointers since it corresponds
to the last page of addressable memory.