nolibc: implement brk() and sbrk()
Be careful, both are implemented on top of the real brk() call which in
fact returns the current end when called with 0 in argument. It's likely
that the system will truncate the memory space to pages and that sbrk()
will be unable to allocate small areas without some keeping an internal
copy of the real end. But for most common use cases it will be enough.