nolibc: fix the stat() syscall
It appears that the output format of the syscall significantly differs
from the one provided by the libc, and often is much larger, explaining
the segfault at boot on MIPS. Depending on the archs, between 1 and 3
syscalls are available (oldstat, newstat, stat64) and none of them is
common to all archs. The formats have to be specified for each arch and
the output has to be recopied, which is a pain. This *seems* to work
though it's possible some issues remain.