nolibc: emit a detailed error when missing alternate syscall number definitions
Some syscalls can be implemented from different __NR_* variants. For
example, sys_dup2() can be implemented based on __NR_dup3 or __NR_dup2.
In this case it's useful to produce an error message when neither are
detected, mentioning the supported alternatives. In case of trouble
with includes, it will help the user search for the right one (e.g
__NR_dup3) instead of just the fallback (__NR_dup2) which might simply
not exist on the platform.

Thanks to Mark Rutland for this suggestion.