nolibc/types: split syscall-specific definitions into their own files
The macros and type definitions used by a number of syscalls were moved
to nolibc-types.h where they will be easier to maintain. A few of them
are arch-specific and must not be moved there (e.g. O_*, sys_stat_struct).
A warning about them was placed at the top of the file.