nolibc/types: move makedev to nolibc-types.h and make it a macro
The makedev() man page says it's supposed to be a macro and that some
OSes have it with the other ones in sys/types.h so it now makes sense
to move it to nolibc-types.h as a macro. Let's also define major() and
minor() that perform the reverse operation.