nolibc: add new "install_khdr_all" and "install_all" targets
These targets are used to install either kernel headers or both kernel
and nolibc headers, for all supported archs together. The arch-specific
asm/ subdirs are renamed to asm-arch-$arch, and asm/ is rebuilt from all
these files in order to include the right one depending on the build
architecture. This is done by reusing a template file (asm-template.h)
for each file found under asm-arch-*, and including the right sub-dir
depending on the current arch.

This allows to use a single unified sysroot for all archs, and to only
change the compiler or the target architecture. This way, a complete
sysroot is much easier to use (a single directory is needed) and much
smaller.

Signed-off-by: Willy Tarreau <w@1wt.eu>